transfered components info and steps to hwd slide
