+++ src/components/Slides/HowWeDo/StepsContent/style.scss
@@ -1,45 +1,44 @@
 .steps-style {
-  .steps-vertical .step:before {
-    height: 100%;
-    width: 0.2rem;
-    --tw-translate-y: -50%;
-    --tw-translate-x: -50%;
-    transform: translate(var(--tw-translate-x), var(--tw-translate-y))
-      rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y))
-      scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y));
-    margin-left: 50%;
-  }
-  .steps-vertical .step:before {
-    width: 0.2rem;
-  }
-  .steps .step:after {
-    font-size: large;
-    font-weight: bolder;
-    height: 2.5rem;
-    width: 2.5rem;
-  }
-  .steps .step::before {
-    background-color: royalblue;
-  }
-  &_list {
-    height: auto;
-  }
 
-  &_text {
-    text-align: start;
-    align-self: flex-end;
-    margin-top: 3rem;
-  }
+    .steps-vertical .step:before {
+        width: 0.2rem;
+    }
 
-  &_title {
-    color: var(--black);
-    font-weight: 900;
-    font-size: 1.2rem;
-  }
+    .steps .step:after {
+        font-size: large;
+        font-weight: bolder;
+        height: 2.5rem;
+        width: 2.5rem;
+    }
 
-  &_desc,
-  p {
-    width: 25rem;
-    text-align: start;
-  }
+    
+
+    &_list {
+        height: auto;
+    }
+
+    &_text {
+        text-align: start;
+        align-self: flex-end;
+        margin-top: 3rem;
+        text-transform: none;
+    }
+
+    &_title {
+        color: var(--black);
+        font-weight: 900;
+        font-size: 1.3rem;
+        padding-bottom: .1rem;
+
+    }
+
+    &_desc,
+    p {
+        // display: none;
+        width: 25rem;
+        font-weight: 500;
+        text-align: start;
+        font-size: 1rem;
+
+    }
 }
