+++ src/components/Slides/HowWeDo/index.tsx
@@ -3,15 +3,20 @@ import SlideContainer from '../SlideContainer';
 import InfoContent from './InfoContent';
 import StepsContent from './StepsContent';
 
-export default function WhatWeDo() {
+export default function HowWeDo() {
   return (
     <SlideContainer>
-      <div className="howWeDoStyle w-screen h-screen">
-        <div className="wwd-style_content wwd-style_info">
-          <InfoContent />
-        </div>
-        <div className="wwd-style_content wwd-style_info">
-          <StepsContent />
+      <div className="w-full">
+        <div className="hwd-style h-[100vh] w-screen">
+          <div className="hwd-style_flex">
+            <div className="hwd-style_content hwd-style_info">
+              <InfoContent />
+            </div>
+
+            <div className="hwd-style_content hwd-style_steps">
+              <StepsContent />
+            </div>
+          </div>
         </div>
       </div>
     </SlideContainer>
