--- src/components/Slides/WhatWeDo/InfoContent/style.scss
@@ -1,30 +1,0 @@
-.content-style {
-
-
-    &_container {
-        flex-direction: column;
-        margin-top: 2rem;
-
-
-        .content-style_text {
-            font-weight: bolder;
-            font-size: 4rem;
-            color: var(--black);
-            text-transform: none;
-        }
-
-        .content-style_subtext {
-            padding-block: 2rem;
-            font-size: 1rem;
-            font-weight: 600;
-
-
-        }
-
-        .content-style_img {
-            height: 90%;
-            width: 90%;
-
-        }
-    }
-}
