--- src/components/Slides/WhatWeDo/StepsContent/style.scss
@@ -1,47 +1,0 @@
-.steps-style {
-
-    .steps-vertical .step:before {
-        width: 0.2rem;
-    }
-
-    .steps .step:after {
-        font-size: large;
-        font-weight: bolder;
-        height: 2.5rem;
-        width: 2.5rem;
-    }
-
-    .steps .step::before {
-        background-color: royalblue;
-    }
-
-
-    &_list {
-        height: auto;
-    }
-
-    &_text {
-        text-align: start;
-        align-self: flex-end;
-        margin-top: 3rem;
-        text-transform: none;
-    }
-
-    &_title {
-        color: var(--black);
-        font-weight: 900;
-        font-size: 1.3rem;
-        padding-bottom: .1rem;
-
-    }
-
-    &_desc,
-    p {
-        // display: none;
-        width: 25rem;
-        font-weight: 500;
-        text-align: start;
-        font-size: 1rem;
-
-    }
-}
