+++ src/components/Slides/WhatWeDo/index.tsx
@@ -1,21 +1,15 @@
-import ContentInfo from './InfoContent';
-import ContentSteps from './StepsContent';
-import SlideContainer from '../SlideContainer';
 import './style.scss';
+import SlideContainer from '../SlideContainer';
 
 export default function WhatWeDo() {
   return (
     <SlideContainer>
-      <div className="w-full">
-        <div className="wwd-style h-[100vh] w-screen">
-          <div className="wwd-style_flex">
-            <div className="wwd-style_content wwd-style_info">
-              <ContentInfo />
-            </div>
-
-            <div className="wwd-style_content wwd-style_steps">
-              <ContentSteps />
-            </div>
+      <div className="hero herostyle">
+        <div className="hero-content text_center">
+          <div className="herostyle_content">
+            <h1 className="herostyle_text">
+              We grow brands with digital <br /> & creative precision
+            </h1>
           </div>
         </div>
       </div>
